% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_download_car_shp.R
\name{batch_download_car_shp}
\alias{batch_download_car_shp}
\title{Função para baixar shapefiles de vários imóveis CAR}
\usage{
batch_download_car_shp(df, coluna_codigo, diretorio_destino = NULL)
}
\arguments{
\item{df}{Dataframe contendo os códigos CAR}

\item{coluna_codigo}{Nome da coluna que contém os códigos CAR}

\item{diretorio_destino}{Diretório onde os shapefiles serão salvos (opcional). Se não informado, será criado 'data/shapefile'.}
}
\description{
Função para baixar shapefiles de vários imóveis CAR
}
