% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_extract_demonstrativo.R
\name{batch_extract_demonstrativo}
\alias{batch_extract_demonstrativo}
\title{Função para processar vários imóveis com base em uma lista de CARs}
\usage{
batch_extract_demonstrativo(car_codes)
}
\arguments{
\item{car_codes}{Vetor de códigos CAR}
}
\value{
Dataframe consolidado com os dados de todos os imóveis
}
\description{
Função para processar vários imóveis com base em uma lista de CARs
}
