% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_sicar_search_cpf.R
\name{batch_sicar_search}
\alias{batch_sicar_search}
\title{Função para aplicar consulta de imóveis a uma lista de CPFs}
\usage{
batch_sicar_search(cpf_list)
}
\arguments{
\item{cpf_list}{Lista de CPFs a serem consultados.}
}
\value{
Dataframe com os resultados das consultas de todos os CPFs.
}
\description{
Função para aplicar consulta de imóveis a uma lista de CPFs
}
