% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_sicar_search_car_code.R
\name{batch_sicar_search_car_code}
\alias{batch_sicar_search_car_code}
\title{Função para consultar imóveis em lote vinculados a uma lista de códigos CAR}
\usage{
batch_sicar_search_car_code(car_codes)
}
\arguments{
\item{car_codes}{Vetor de códigos CAR a serem consultados.}
}
\value{
Dataframe com os resultados das consultas de todos os códigos CAR.
}
\description{
Função para consultar imóveis em lote vinculados a uma lista de códigos CAR
}
