% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_consultar_dominialidade.R
\name{consultar_dominialidade_varios_car}
\alias{consultar_dominialidade_varios_car}
\title{Função para consultar a dominialidade de vários imóveis CAR}
\usage{
consultar_dominialidade_varios_car(codigos_car, pb = NULL)
}
\arguments{
\item{codigos_car}{Vetor de códigos CAR para consulta}

\item{pb}{(opcional) Barra de progresso da consulta.}
}
\value{
Um dataframe consolidado com as informações de dominialidade (nome, CPF/CNPJ, tipo, código CAR e ID).
}
\description{
Função para consultar a dominialidade de vários imóveis CAR
}
