% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicar_extract_demonstrativo.R
\name{extrair_dados_demonstrativo}
\alias{extrair_dados_demonstrativo}
\title{Função para extrair o texto de uma div com base no id e na classe}
\usage{
extrair_dados_demonstrativo(page, div_id, class)
}
\arguments{
\item{page}{HTML da página do imóvel}

\item{div_id}{ID da div que contém o dado desejado}

\item{class}{Classe do HTML onde os dados estão}
}
\value{
Texto extraído ou NA caso o dado não seja encontrado
}
\description{
Função para extrair o texto de uma div com base no id e na classe
}
