% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicar_search_car_code.R
\name{formatar_string}
\alias{formatar_string}
\title{Função para formatar a string do código CAR}
\usage{
formatar_string(input_string)
}
\arguments{
\item{input_string}{String de código CAR a ser formatada.}
}
\value{
Retorna o código CAR formatado com pontos.
}
\description{
Função para formatar a string do código CAR
}
