% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicar_consultar_dominialidade.R
\name{sicar_consultar_dominialidade}
\alias{sicar_consultar_dominialidade}
\title{Função para consultar a dominialidade de um imóvel CAR a partir do código CAR}
\usage{
sicar_consultar_dominialidade(codigo_car, pb = NULL)
}
\arguments{
\item{codigo_car}{Código CAR para consulta}

\item{pb}{(opcional) Barra de progresso da consulta.}
}
\value{
Um dataframe com as informações de dominialidade (nome, CPF/CNPJ, tipo, código CAR e ID).
}
\description{
Função para consultar a dominialidade de um imóvel CAR a partir do código CAR
}
