% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicar_download_car_shp.R
\name{sicar_download_car_shp}
\alias{sicar_download_car_shp}
\title{Função para baixar o shapefile de um imóvel CAR}
\usage{
sicar_download_car_shp(codigo_imovel, diretorio_destino = NULL)
}
\arguments{
\item{codigo_imovel}{Código CAR do imóvel a ser baixado}

\item{diretorio_destino}{Diretório onde o shapefile será salvo (opcional). Se não informado, será criado 'data/shapefile'.}
}
\description{
Função para baixar o shapefile de um imóvel CAR
}
