% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicar_extract_demonstrativo.R
\name{sicar_extract_demonstrativo}
\alias{sicar_extract_demonstrativo}
\title{Função para processar um imóvel pelo código CAR e extrair dados do demonstrativo}
\usage{
sicar_extract_demonstrativo(car_code)
}
\arguments{
\item{car_code}{Código CAR do imóvel a ser processado}
}
\value{
Dataframe com os dados extraídos do imóvel
}
\description{
Função para processar um imóvel pelo código CAR e extrair dados do demonstrativo
}
