% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicar_search_car_code.R
\name{sicar_search_car_code}
\alias{sicar_search_car_code}
\title{Função para consultar imóveis vinculados a um código CAR no SICAR}
\usage{
sicar_search_car_code(codigo, pb = NULL)
}
\arguments{
\item{codigo}{Código CAR para consulta.}

\item{pb}{(opcional) Barra de progresso da consulta.}
}
\value{
Um dataframe com os dados de imóveis vinculados ao código CAR.
}
\description{
Função para consultar imóveis vinculados a um código CAR no SICAR
}
