% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicar_search.R
\name{sicar_search_cpf}
\alias{sicar_search_cpf}
\title{Função para consultar imóveis vinculados a um CPF no sistema SICAR}
\usage{
sicar_search_cpf(cpf)
}
\arguments{
\item{cpf}{Número do CPF para consulta (11 dígitos).}
}
\value{
Um dataframe com os dados de imóveis vinculados ao CPF.
}
\description{
Função para consultar imóveis vinculados a um CPF no sistema SICAR
}
